configuring data for the api in connection,station and train

+++ GetDelaysAPI/StationData/ArrivalDeparture.cs
@@ -9,7 +9,7 @@ namespace GetDelaysAPI
     public class ArrivalDeparture
     {
         public int id { get; set; }
-        public int delay { get; set; }
+        public double delay { get; set; }
         public string station { get; set; }
         public string vehicle { get; set; }
         public string time { get; set; }
